+++ common/src/main/java/fr/tathan/sky_aesthetics/client/skies/renderer/SkyRenderer.java
@@ -21,6 +21,7 @@ import net.minecraft.client.server.IntegratedServer;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.phys.Vec3;
 import org.joml.Matrix4f;
+import org.joml.Vector4f;
 
 import java.util.*;
 
@@ -34,7 +35,7 @@ public class SkyRenderer {
         this.properties = properties;
 
         if(!properties.stars().vanilla()) {
-            starBuffer = StarHelper.createStars(properties.stars().scale(), properties.stars().count(), properties.stars().color().r(), properties.stars().color().g(), properties.stars().color().b(), properties.constellations());
+            starBuffer = StarHelper.createStars(properties.stars().scale(), properties.stars().count(), (int) properties.stars().color().x(), (int) properties.stars().color().y(), (int) properties.stars().color().z(), properties.constellations());
         } else {
             starBuffer = StarHelper.createVanillaStars();
         }
@@ -55,23 +56,18 @@ public class SkyRenderer {
         float nightAngle = dayAngle + 180;
 
         Vec3 vec3 = level.getSkyColor(camera.getPosition(), partialTick);
+        Vector4f vec4 = new Vector4f((float) vec3.x,(float) vec3.y,(float) vec3.z, 1.0f);
 
         if(properties.skyColor().customColor()) {
-            vec3 = properties.skyColor().color();
+            vec4 = properties.skyColor().color();
         }
-
-        float r = (float) vec3.x;
-        float g = (float) vec3.y;
-        float b = (float) vec3.z;
 
         FogRenderer.levelFogColor();
         RenderSystem.depthMask(false);
 
-        RenderSystem.setShaderColor(r, g, b, 1.0f);
+        RenderSystem.setShaderColor(vec4.x, vec4.y, vec4.z, vec4.w);
 
         ShaderInstance shaderInstance = RenderSystem.getShader();
-
-
 
         if(Objects.equals(properties.skyType(), "NORMAL")) {
             SkyHelper.drawSky(poseStack.last().pose(), projectionMatrix, shaderInstance);
@@ -84,11 +80,9 @@ public class SkyRenderer {
 
         properties.stars().shootingStars().ifPresent((shootingStar -> handleShootingStars(level, poseStack, projectionMatrix, properties.stars(), partialTick)));
 
-
         // Sun
         if (customVanillaObject.sun()) {
            SkyHelper.drawCelestialBody(customVanillaObject.sunTexture(), tesselator, poseStack, customVanillaObject.sunHeight(), customVanillaObject.sunSize(), dayAngle, true);
-
         }
 
         // Moon