+++ changelog.md
@@ -1,4 +1,13 @@
-# Sky Aesthetics 1.2.13
+# Sky Aesthetics 1.3.0
+
+## Breaking Changes
+- Change the color of the stars to a Vector3
+- Change the color of the constellation to a Vector3
+```json
+{
+  "color": [160, 180, -1]
+}
+```
 
 ## Fix
 - Fix sunrise color rendering ar midnigh (#14)